fix purge

+++ docker-stacks/00-network/hooks/01-remove-data.sh
@@ -1,3 +1,9 @@
-rm "${STACK_HOME}/.initialized"
-rm "${STACK_HOME}/.externaldomain"
-rm "${STACK_HOME}/.version"
+if [[ -e "${STACK_HOME}/.initialized" ]] then
+    rm "${STACK_HOME}/.initialized"
+fi
+if [[ -e "${STACK_HOME}/.externaldomain" ]] then
+    rm "${STACK_HOME}/.externaldomain"
+fi 
+if [[ -e "${STACK_HOME}/.version" ]] then
+    rm "${STACK_HOME}/.version"
+fi
