+++ docker-stacks/02-keycloak/etc/hooks/purge.d/01-remove-data-and-config.sh
@@ -52,6 +52,12 @@ if [[ -e "${SIMVA_CONFIG_HOME}/keycloak/realm-data.prod.yml" ]]; then
     rm "${SIMVA_CONFIG_HOME}/keycloak/realm-data.prod.yml"
 fi
 
-rm "${STACK_HOME}/.initialized"
-rm "${STACK_HOME}/.version"
-rm "${STACK_HOME}/.externaldomain"
+if [[ -e "${STACK_HOME}/.initialized" ]] then
+    rm "${STACK_HOME}/.initialized"
+fi
+if [[ -e "${STACK_HOME}/.externaldomain" ]] then
+    rm "${STACK_HOME}/.externaldomain"
+fi 
+if [[ -e "${STACK_HOME}/.version" ]] then
+    rm "${STACK_HOME}/.version"
+fi
