+++ docker-stacks/05-kafka/etc/hooks/purge.d/01-remove-data-and-config.sh
@@ -36,6 +36,12 @@ rm -rf ./*
 cd "${SIMVA_DATA_HOME}/kafka/data/zoo1"
 rm -rf ./*
 
-rm "${STACK_HOME}/.initialized"
-rm "${STACK_HOME}/.version"
-rm "${STACK_HOME}/.externaldomain"
+if [[ -e "${STACK_HOME}/.initialized" ]] then
+    rm "${STACK_HOME}/.initialized"
+fi
+if [[ -e "${STACK_HOME}/.externaldomain" ]] then
+    rm "${STACK_HOME}/.externaldomain"
+fi 
+if [[ -e "${STACK_HOME}/.version" ]] then
+    rm "${STACK_HOME}/.version"
+fi
