+++ docker-stacks/06-simva/etc/hooks/purge.d/01-remove-data-and-config.sh
@@ -22,6 +22,12 @@ if [[ -e "${SIMVA_PROJECT_DIR}/trace-allocator.log" ]]; then
     rm "${SIMVA_PROJECT_DIR}/trace-allocator.log"
 fi
 
-rm "${STACK_HOME}/.initialized"
-rm "${STACK_HOME}/.version"
-rm "${STACK_HOME}/.externaldomain"
+if [[ -e "${STACK_HOME}/.initialized" ]] then
+    rm "${STACK_HOME}/.initialized"
+fi
+if [[ -e "${STACK_HOME}/.externaldomain" ]] then
+    rm "${STACK_HOME}/.externaldomain"
+fi 
+if [[ -e "${STACK_HOME}/.version" ]] then
+    rm "${STACK_HOME}/.version"
+fi
